+++ is-eurail-cheaper/app/css/globals.scss
@@ -357,10 +357,15 @@ body {
   margin-right: 4px;
 }
 
+.inline-block {
+  display: inline-block;
+}
+
 .city {
   border-radius: 25px;
-  padding: 0 5px;
+  padding-left: 5px;
   border: 1px solid lightgray;
+  overflow: hidden;
   //transition: all 0.1s;
   //border: 1px solid red;
 }
@@ -370,22 +375,45 @@ body {
   //transition: all 0.1s;
 }
 
+.dot, .city-name {
+  //padding-bottom: 5px;
+}
+
 .edit {
-  width: 18px;
+  width: 22px;
   height: 18px;
-  padding-left: 4px;
-  padding-top: 4px;
-  display: inline-block;
   //filter: brightness(85%);
+  padding-left: 1px;
+  margin-top: 3px;
   opacity: 0.2;
   //transition: all 0.2s;
 }
 
 .bright {
   opacity: 1 !important;
+  //filter: invert(1);
   //transition: all 0.1s;
 }
 
 .action-tag:hover {
   filter: brightness(85%);
+}
+
+.edit-container {
+  margin-left: 4px;
+  //padding-top: 4px;
+  width: 24px;
+  display: inline-block;
+}
+
+.edit-container:hover {
+  background: #293045;
+}
+
+.edit-container:hover .edit {
+  filter: invert(1);
+}
+
+.city-name-input {
+  display: inline-block;
 }
